Update MeshCore_Interface.py
Figured out that issues on RPi 5 were being caused by MeshChat, NomadNet, and rBrowser conflicts around config file updates. Fixed by making rnsd the core service that stands up the RNS stack, and then MeshChat, NomadNet, and rBrowser all start as services with rnsd.service as a requirement.

This version of the interface has a line that isn't valid to try to ensure the MeshCore node is still available, so that's commented out while I figure out an alternative. I'm also getting an error where my targeted endpoint is causing an error about it being a string when it should be a hex key.